requirements.txt
asgiref==3.3.1
cycler==0.10.0
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
fpdf==1.7.2
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
